Weird bug.
There is a weird bug (probably in selector but I don't know.) where the
order of the things appearing in the urls.txt file is messing stuff up.

"It doesn't work."  I know I know, I'm just noting it for myself.

I bet it has to do with the scope of the '@wrap' directive.